feat: change prep and filter behaviour

+++ src/conchord/preprocess.py
@@ -16,7 +16,7 @@ from src.conchord.utils.parser import get_preprocess_parser
 from src.conchord.utils.utils import align_labels_to_frames, convert_arff_to_lab, load_lab_file
 
 config = Config()
-BASE_DIR = Path(__file__).resolve().parents[2] / 'data' / 'preprocessed'
+PREPROCESS_DIR = Path(__file__).resolve().parents[2] / 'data' / 'preprocessed'
 FRAME_DURATION = config.AUDIO_PARAMS['hop_length'] / config.AUDIO_PARAMS['sample_rate']
 SEQUENCE_DURATION = 1  # second
 SEQUENCE_LENGTH = int(np.ceil(SEQUENCE_DURATION / FRAME_DURATION))  # 22 frames per second
@@ -66,8 +66,8 @@ def _validate_instruments(AAM_instruments: list[str]) -> bool:
     return True
 
 
-def _load_dataset(dataset_name: str) -> dict:
-    path = Path(BASE_DIR) / f'{dataset_name}.npz'
+def _load_dataset_to_filter(dataset_name: str) -> dict:
+    path = Path(PREPROCESS_DIR) / f'{dataset_name}.npz'
     if not path.exists():
         logging.error(f'{dataset_name} not found at {path}')
         return {}
@@ -129,7 +129,12 @@ def add_chroma_noise(
 
 
 def _preprocess_idmt_dataset(
-    dataset: dict, src_dir: Path, output_path: Path, hop_len: int, sample_rate: int, seq_len: int = SEQUENCE_LENGTH
+    dataset: dict,
+    src_dir: Path,
+    output_path: Path,
+    hop_len: int,
+    sample_rate: int,
+    seq_len: int = SEQUENCE_LENGTH,
 ) -> None:
     """
     Processes the IDMT dataset by:
@@ -198,6 +203,7 @@ def _preprocess_aam_dataset(
     - Loads MIDI files and extracting chroma features from note velocities.
     - Creates binary note presence matrices across 128 MIDI pitches.
     - Aligns chroma frames with chord labels.
+    - Exclude completely silent bins.
     - Adds random noise to chroma bins.
     - Reshapes frames into sequences.
     Args:
@@ -205,7 +211,7 @@ def _preprocess_aam_dataset(
         general_noise: Normal distribution Z~N(µ,σ)
     """
     X, Y_chords, Y_notes, sources, categories = [], [], [], [], []
-    temp_dir = BASE_DIR / '_temp_AAM'
+    temp_dir = PREPROCESS_DIR / '_temp_AAM'
     temp_dir.mkdir(parents=True, exist_ok=True)
 
     mid_files = sorted(
@@ -227,7 +233,7 @@ def _preprocess_aam_dataset(
         try:
             midi_data = PrettyMIDI(str(midi))  # turn Path object to path as a string``
         except KeySignatureError:
-            tqdm.write(f'[Warning] - Invalid key signature in {midi}, skipping')
+            tqdm.write(f'[Warning] - Invalid key signature in {(midi.name)[:-4]}, skipping')
             continue
 
         lab_path = temp_dir / f'{midi_id}_beatinfo.lab'
@@ -362,36 +368,46 @@ def preprocess_data(
     dataset_names: list[str],
     idmt_hop_len: int,
     idmt_sr: int,
-    silence_noise: list[float, float],  #! list due to the format of parser.
+    silence_noise: list[float, float],  #! type list due to the format of parser.
     general_noise: list[float, float],
+    prep_idmt: bool = False,
+    prep_aam: bool = False,
+    prep_maestro: bool = False,
 ) -> None:
     """
     Main entry point for preprocessing datasets.
     - Iterates through the provided dataset names and triggers the appropriate
     - processes function depending on the dataset type.
-    - Skips processing if the output .npz file already exists.
+    - Skips processing if the output .npz file already exists and skipping is enabled.
     """
     for name in dataset_names:
+        prep_msg = f'Skipping {name} – already preprocessed.'
         src = Path(__file__).resolve().parents[2] / 'data' / 'datasets' / name
-        out = BASE_DIR / f'{name}.npz'
+        out = PREPROCESS_DIR / f'{name}.npz'
         if not src.exists():
             logging.error(f'{src} not found!')
             return
-        if out.exists():
-            logging.info(f'Skipping {name} – already preprocessed.')
-            continue
 
-        logging.info(f'{name} preprocessing started')
+        logging.info(f'{name} preprocessing initiated:')
         if name == 'IDMT':
+            if out.exists() and prep_idmt:
+                logging.info(prep_msg)
+                continue
             _preprocess_idmt_dataset(config.DATASETS[name], src, out, hop_len=idmt_hop_len, sample_rate=idmt_sr)
         elif name == 'AAM':
+            if out.exists() and prep_aam:
+                logging.info(prep_msg)
+                continue
             try:  #! change noise args to tuple
                 _preprocess_aam_dataset(src, out, tuple(silence_noise), tuple(general_noise))
             except KeyboardInterrupt:
                 logging.warning('Interrupted. Cleaning up...')
-                shutil.rmtree(BASE_DIR / '_temp_AAM', ignore_errors=True)
+                shutil.rmtree(PREPROCESS_DIR / '_temp_AAM', ignore_errors=True)
                 raise
         elif name == 'MAESTRO':
+            if out.exists() and prep_maestro:
+                logging.info(prep_msg)
+                continue
             _preprocess_maestro_dataset(config.DATASETS[name], src, out, tuple(silence_noise), tuple(general_noise))
 
     logging.info('Preprocessing finished.')
@@ -407,15 +423,20 @@ def _filter_idmt_dataset(
     IDMT_size: int,
     guitar_ratio: float,
     output_path: Path,
-    use_max_size: bool = False,
+    size_tracker: dict[str, int],
+    use_max_size: bool = True,
 ) -> None:
     """
     Creates a filtered subset of the IDMT dataset or copy full dataset if use_max_size is True.
     """
+    logging.info('IDMT filterling initiated:')
     if use_max_size:
-        source_path = Path(BASE_DIR) / 'IDMT.npz'
-        shutil.copy(source_path, output_path)
-        logging.info(f'Copied full IDMT dataset to {output_path}')
+        source_path = Path(PREPROCESS_DIR) / 'IDMT.npz'
+        if source_path.exists():
+            shutil.copy(source_path, output_path)
+            logging.info(f'Max size - Copied full IDMT dataset to {output_path}')
+        else:
+            logging.error(f'Source not found: {source_path}')
         return
 
     X = data['X']
@@ -441,6 +462,7 @@ def _filter_idmt_dataset(
     selected_indices = np.concatenate([guitar_indices, nonguitar_indices])
     rng.shuffle(selected_indices)
 
+    size_tracker['IDMT'] = len(X[selected_indices])
     save_npz(
         output_path,
         X=X[selected_indices],
@@ -455,13 +477,15 @@ def _filter_aam_dataset(
     AAM_size: int,
     AAM_instruments: list[str],
     output_path: Path,
-    use_max_size: bool = False,
-    use_all_aam_instruments: bool = False,
+    size_tracker: dict[str, int],
+    use_max_size: bool = True,
+    use_all_AAM_instruments: bool = False,
 ) -> None:
     """
     - Creates a filtered subset of AAM with specified instruments.
     - Slices entire subsets of AAM for each instrument if use_max_size is True
     """
+    logging.info('AAM filterling initiated:')
     X = data['X']
     Y_chords = data['Y_chords']
     Y_notes = data['Y_notes']
@@ -471,7 +495,7 @@ def _filter_aam_dataset(
     rng = np.random.default_rng(config.SEED)
     selected_indices = []
 
-    if use_all_aam_instruments:
+    if use_all_AAM_instruments:
         selected_indices = np.arange(len(X))
         if not use_max_size:
             if AAM_size > len(selected_indices):
@@ -480,7 +504,6 @@ def _filter_aam_dataset(
         rng.shuffle(selected_indices)
 
     else:
-        selected_indices = []
         for inst in AAM_instruments:
             mask = categories == inst
             indices = np.where(mask)[0]
@@ -493,10 +516,13 @@ def _filter_aam_dataset(
                         f'Requested {samples_per_instrument} samples for {inst}, but only {available} available.'
                     )
                 indices = rng.choice(indices, size=samples_per_instrument, replace=False)
-
             selected_indices.extend(indices)
         rng.shuffle(selected_indices)
 
+    if use_max_size:
+        logging.info(f'Max size - Copied full AAM dataset to {output_path}')
+
+    size_tracker['AAM'] = len(X[selected_indices])
     save_npz(
         output_path,
         X=X[selected_indices],
@@ -507,14 +533,17 @@ def _filter_aam_dataset(
     )
 
 
-def _filter_maestro_dataset(data: dict, MAESTRO_size: int, output_path: Path, use_max_size: bool = False) -> None:
+def _filter_maestro_dataset(
+    data: dict, MAESTRO_size: int, output_path: Path, size_tracker: dict[str, int], use_max_size: bool = True
+) -> None:
     """
     Creates a filtered subset of MAESTRO or copy full dataset if use_max_size is True
     """
+    logging.info('MAESTRO filterling initiated:')
     if use_max_size:
-        source_path = BASE_DIR / 'MAESTRO.npz'
+        source_path = PREPROCESS_DIR / 'MAESTRO.npz'
         shutil.copy(source_path, output_path)
-        logging.info(f'Copied full MAESTRO dataset to {output_path}')
+        logging.info(f'Max size - Copied full MAESTRO dataset to {output_path}')
         return
 
     X = data['X']
@@ -527,6 +556,7 @@ def _filter_maestro_dataset(data: dict, MAESTRO_size: int, output_path: Path, us
     selected_indices = rng.choice(indices, size=MAESTRO_size, replace=False)
     rng.shuffle(selected_indices)
 
+    size_tracker['MAESTRO'] = len(X[selected_indices])
     save_npz(
         output_path,
         X=X[selected_indices],
@@ -544,9 +574,13 @@ def filter_data(
     MAESTRO_ratio: float,
     IDMT_guitar_ratio: float,
     AAM_instruments: list[str],
-    use_max_size: bool = False,
-    use_all_aam_instruments: bool = False,
+    use_max_size: bool = True,
+    use_all_instruments: bool = False,
+    filter_idmt: bool = False,
+    filter_aam: bool = False,
+    filter_maestro: bool = False,
 ):
+    logging.info('Data filtering/slicing initiated:')
     (Path(__file__).resolve().parents[2] / 'data' / 'filtered').mkdir(parents=True, exist_ok=True)
 
     dataset_split_ratios = {'IDMT_ratio': IDMT_ratio, 'AAM_ratio': AAM_ratio, 'MAESTRO_ratio': MAESTRO_ratio}
@@ -556,21 +590,40 @@ def filter_data(
         return
 
     IDMT_size, AAM_size, MAESTRO_size = [int(np.floor(filter_size * ratio)) for ratio in dataset_split_ratios.values()]
-
+    sizes = {'IDMT': IDMT_size, 'AAM': AAM_size, 'MAESTRO': MAESTRO_size}
+    fliter_flags = {
+        'IDMT': filter_idmt,
+        'AAM': filter_aam,
+        'MAESTRO': filter_maestro,
+    }
     for dataset_name in dataset_names:
         output_path = Path(__file__).resolve().parents[2] / 'data' / 'filtered' / f'{dataset_name}.npz'
-        data = _load_dataset(dataset_name)
+        if output_path.exists() and fliter_flags.get(dataset_name, False):
+            logging.info(f'Skipping {dataset_name} — filtered version already exists.')
+            continue
+
+        data = _load_dataset_to_filter(dataset_name)
         if not data:
             return
 
         if dataset_name == 'IDMT':
-            _filter_idmt_dataset(data, IDMT_size, IDMT_guitar_ratio, output_path, use_max_size)
+            _filter_idmt_dataset(
+                data, IDMT_size, IDMT_guitar_ratio, output_path, use_max_size=use_max_size, size_tracker=sizes
+            )
         elif dataset_name == 'AAM':
-            _filter_aam_dataset(data, AAM_size, AAM_instruments, output_path, use_max_size, use_all_aam_instruments)
+            _filter_aam_dataset(
+                data,
+                AAM_size,
+                AAM_instruments,
+                output_path,
+                use_max_size=use_max_size,
+                use_all_AAM_instruments=use_all_instruments,
+                size_tracker=sizes,
+            )
         elif dataset_name == 'MAESTRO':
-            _filter_maestro_dataset(data, MAESTRO_size, output_path, use_max_size)
-    logging.info(f'IDMT_size: {IDMT_size:,} | AAM_size: {AAM_size:,} | MAESTRO_size: {MAESTRO_size:,}')
+            _filter_maestro_dataset(data, MAESTRO_size, output_path, use_max_size=use_max_size, size_tracker=sizes)
     logging.info('Data filtering/slicing complete.')
+    logging.info(f'IDMT size: {sizes["IDMT"]:,} | AAM size: {sizes["AAM"]:,} | MAESTRO size: {sizes["MAESTRO"]:,}')
 
 
 # ============================================
@@ -584,8 +637,9 @@ def stack_datasets(
     """
     Loads and stacks previously filtered datasets.
     """
+    logging.info('Stackig and organizig filtered data...')
     datasets_dir = Path(__file__).resolve().parents[2] / 'data' / 'filtered'
-    all_X, all_Y_chords, all_Y_notes, all_sources = [], [], [], []
+    all_X, all_Y_chords, all_Y_notes, all_sources, all_datasets = [], [], [], [], []
 
     for dataset in datasets:
         ds_path = datasets_dir / f'{dataset}.npz'
@@ -593,6 +647,7 @@ def stack_datasets(
         X = data['X']
         all_X.append(X)
         all_sources.extend(data['sources'])
+        all_datasets.extend([dataset] * len(data['sources']))
 
         if 'Y_notes' in data:
             all_Y_notes.append(data['Y_notes'])
@@ -608,6 +663,7 @@ def stack_datasets(
     Y_notes = np.vstack(all_Y_notes)
     Y_chords = np.concatenate(all_Y_chords, axis=0)
     sources = np.array(all_sources)
+    dataset_names = np.array(all_datasets)
 
     # Filter out samples with missing chord labels or all NaN notes
     chord_mask = Y_chords[:, 0] != 'MISSING'  # Only the first frame suffices
@@ -616,20 +672,24 @@ def stack_datasets(
     X_chords = X[chord_mask]
     Y_chords_filtered = Y_chords[chord_mask]
     sources_chords = sources[chord_mask]
+    datasets_chords = dataset_names[chord_mask]
     X_notes = X[note_mask]
     Y_notes_filtered = Y_notes[note_mask]
     sources_notes = sources[note_mask]
+    datasets_notes = dataset_names[note_mask]
 
     splitted_data = {
         'chords': {
             'X': X_chords,
             'Y': Y_chords_filtered,
             'sources': sources_chords,
+            'datasets': datasets_chords,
         },
         'notes': {
             'X': X_notes,
             'Y': Y_notes_filtered,
             'sources': sources_notes,
+            'datasets': datasets_notes,
         },
     }
     return splitted_data
@@ -641,6 +701,7 @@ def split_data(
     """
     Splits the previously stacked datasets into test, train and validation sets.
     """
+    logging.info(f'Splitting {label_type.upper()} data into test, train, val sets...')
     split_map_ratios = {
         'train': train_ratio,
         'val': val_ratio,
@@ -655,6 +716,7 @@ def split_data(
     X = data[label_type]['X']
     Y = data[label_type]['Y']
     sources = data[label_type]['sources']
+    datasets = data[label_type]['datasets']
 
     indices = np.arange(len(X))
     np.random.default_rng(config.SEED).shuffle(indices)
@@ -682,6 +744,7 @@ def split_data(
             X=X[selected_indices],
             Y=Y[selected_indices],
             sources=sources[selected_indices],
+            datasets=datasets[selected_indices]
         )
         indices_lenghts.append(len(selected_indices))
     logging.info(
@@ -698,23 +761,37 @@ def main():
     setup_logger()
     parser = get_preprocess_parser()
     args = parser.parse_args()
+
+    # Resolve force flags from list
+    force_prep_idmt = 'IDMT' not in args.force_preprocess
+    force_prep_aam = 'AAM' not in args.force_preprocess
+    force_prep_maestro = 'MAESTRO' not in args.force_preprocess
     preprocess_data(
         dataset_names=args.datasets,
         idmt_hop_len=args.hop_length,
         idmt_sr=args.sample_rate,
         silence_noise=args.silence_noise,
         general_noise=args.general_noise,
+        prep_idmt=force_prep_idmt,
+        prep_aam=force_prep_aam,
+        prep_maestro=force_prep_maestro,
     )
+    force_filter_idmt = 'IDMT' not in args.force_filter
+    force_filter_aam = 'AAM' not in args.force_filter
+    force_filter_maestro = 'MAESTRO' not in args.force_filter
     filter_data(
         dataset_names=args.datasets,
         filter_size=args.filter_size,
         use_max_size=args.use_max_size,
-        use_all_aam_instruments=args.use_all_aam_instruments,
+        use_all_instruments=args.use_all_aam_instruments,
         IDMT_ratio=args.idmt_ratio,
         AAM_ratio=args.aam_ratio,
         MAESTRO_ratio=args.maestro_ratio,
         IDMT_guitar_ratio=args.idmt_guitar_ratio,
         AAM_instruments=args.aam_instruments,
+        filter_idmt=force_filter_idmt,
+        filter_aam=force_filter_aam,
+        filter_maestro=force_filter_maestro,
     )
     data = stack_datasets(datasets=args.datasets)
     for type in ['notes', 'chords']:
